+++ bpf/libbpf/ksym_helpers.c
@@ -1,0 +1,1 @@
+../../fs/procfs/kallsyms/ksym_helpers.c